requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
pytesseract
Pillow